move cookie config to config file
